requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0
matplotlib>=3.3.4
higher>=0.2.1
tqdm>=4.56.0
pyyaml>=5.4.1
Pillow>=8.1.0
learn2learn>=0.1.5
